backend/requirements.txt
# backend
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2
Django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-filter
drf-url-filters==0.5.1
drf-spectacular==0.27.2
django-environ>=0.11.0
django-redis==5.4.0
flake8
gunicorn>=21.0.0
Pillow>=10.1.0
psycopg2-binary>=2.9.10
python-dotenv>=1.0.0

# testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
